bugfix: fixed NPE
git-svn-id: https://server.mycore.de/svn/mycore/trunk@8780 7089d063-bc14-496d-ac90-53de1872eb6c

+++ sources/org/mycore/backend/filesystem/MCRCStoreLocalFilesystem.java
@@ -139,12 +139,14 @@ public class MCRCStoreLocalFilesystem extends MCRContentStore {
 
         for (int i = numDirs; i > 0; i--) {
             File dir = new File(baseDir, dirs[i - 1]);
-
-            if (dir.listFiles().length > 0) {
-                break;
+            if (dir.isDirectory()) {
+                // dir is directory and exists
+                // ensures that listFiles below does not throw NPE
+                if (dir.listFiles().length > 0) {
+                    break;
+                }
+                dir.delete();
             }
-
-            dir.delete();
         }
     }
